New C-side prototype.
The new SpiNNaker application expects a series of commands to be loaded which
control how packets are generated and what is recorded. This enormously
simplifies the task of configuring the program and will hopefully result in a
vastly simpler and more flexible Python side.

The Python side needs suitably re-architecting/writing.